68.13.03 - REST - Klasyczny CRUD  za pomoca RESTa - Delete_DELETE (usuwanie bledu zaleznosci jednego serwisu od drugiego poprzez uzycie @Lazy oraz zabezpieczanie błędu kasowania Roomu, który juz jest w rezerwacji

+++ src/main/java/pl/clockworkjava/gnomix/domain/room/RoomService.java
@@ -1,69 +1,92 @@
 package pl.clockworkjava.gnomix.domain.room;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
+import pl.clockworkjava.gnomix.domain.reservation.Reservation;
+import pl.clockworkjava.gnomix.domain.reservation.ReservationService;
 
-import java.util.*;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
 @Service
 public class RoomService {
 
-
     private RoomRepository repository;
+    private ReservationService reservationService;
 
     @Autowired
-    public RoomService(RoomRepository repository) {
+    public RoomService(RoomRepository repository, ReservationService reservationService) {
         this.repository = repository;
+        this.reservationService = reservationService;
     }
 
     public List<Room> findAllRooms() {
-        return this.repository.findAll();
+        return repository.findAll();
     }
 
-    public Room createNewRoom(String roomNumber, String bedsDescription, String description, List<String> photosUrl) {
-        List<BedType> beds = getBedTypesList(bedsDescription);
+    public Room createNewRoom(String roomNumber, String bedsDesc, String description, List<String> photosUrls) {
 
-        Room newOne = new Room(roomNumber, beds, description, photosUrl);
+        List<BedType> beds = getBedTypesList(bedsDesc);
+
+        Room newOne = new Room(roomNumber, beds, description, photosUrls);
 
         return this.repository.save(newOne);
     }
-    public Room createNewRoom(String roomNumber, List<BedType> beds, String description, List<String> photosUrl) {
 
-        Room newOne = new Room(roomNumber, beds, description, photosUrl);
+    public Room createNewRoom(String roomNumber, List<BedType> beds, String description, List<String> photosUrls) {
+
+        Room newOne = new Room(roomNumber, beds, description, photosUrls);
 
         return this.repository.save(newOne);
     }
 
     public void removeRoomById(long id) {
 
+        boolean thereIsReservationForThisRoom = this.reservationService
+                .getAll()
+                .stream()
+                .anyMatch(r -> r.getRoom().getRoomId()==id);
+
+        if(thereIsReservationForThisRoom) {
+            throw new IllegalStateException("There is reservation for this room");
+        }
+
         this.repository.deleteById(id);
+
     }
 
     public Room findById(long id) {
+
         return this.repository.getById(id);
-
     }
 
+    public void update(long id, String number, String bedsDesc, String description, List<String> photosUrls) {
 
-    public void update(long id, String roomNumber, String bedsDescription, String description, List<String> photosUrl) {
-        Room roomToUpdate = this.repository.getById(id);
-        List<BedType> beds = getBedTypesList(bedsDescription);
-        roomToUpdate.update(roomNumber, beds, description, photosUrl);
-        this.repository.save(roomToUpdate);
+        Room toUpdate = this.repository.getById(id);
+
+        List<BedType> beds = getBedTypesList(bedsDesc);
+
+
+        toUpdate.update(number, beds, description, photosUrls);
+
+        this.repository.save(toUpdate);
     }
 
-    private List<BedType> getBedTypesList(String bedsDescription) {
-        String[] splitedBedsDescription = bedsDescription.split("\\+");
-        return Arrays.stream(splitedBedsDescription)
+    private List<BedType> getBedTypesList(String bedsDesc) {
+        String[] splitedBedDec = bedsDesc.split("\\+");
+
+        return Arrays.stream(splitedBedDec)
                 .map(stringToBedTypeMapping)
                 .collect(Collectors.toList());
     }
 
-    private final Function<String, BedType> stringToBedTypeMapping
-            = value -> {
+    private final Function<String, BedType> stringToBedTypeMapping = value -> {
         if ("1".equals(value)) {
             return BedType.SINGLE;
         } else if ("2".equals(value)) {
@@ -74,13 +97,15 @@ public class RoomService {
     };
 
     public List<Room> getRoomsForSize(int size) {
-        if (size <= 0) {
+
+        if(size <= 0) {
             return new ArrayList<>();
         }
-        return this.repository.findAll()
-                .stream().filter(room -> room.getSize() >= size)
-                .collect(Collectors.toList());
 
+        return this.repository.findAll()
+                .stream()
+                .filter( r -> r.getSize()>=size)
+                .collect(Collectors.toList());
     }
 
     public Optional<Room> getRoomById(long roomId) {
